using the backstage discovery api, split components and added tests
